United swaps parameters
