+++ runtime/src/lib.rs
@@ -102,24 +102,15 @@ parameter_types! {
   pub const DisputeFactor: Balance = 2 * BASE;
   pub const DisputePeriod: BlockNumber = DAYS;
   pub const ExistentialDeposit: u128 = EXISTENTIAL_DEPOSIT;
-  pub const ExitFee: Balance = 0;
   pub const GetNativeCurrencyId: Asset<MarketId> = Asset::Ztg;
-  pub const MaxAssets: usize = 9;
   pub const MaxCategories: u16 = 8;
   pub const MaxDisputes: u16 = 6;
-  pub const MaxInRatio: Balance = BASE / 2;
   pub const MaxLocks: u32 = 50;
-  pub const MaxOutRatio: Balance = (BASE / 3) + 1;
-  pub const MaxTotalWeight: Balance = 50 * BASE;
-  pub const MaxWeight: Balance = 50 * BASE;
   pub const MinimumPeriod: u64 = SLOT_DURATION / 2;
-  pub const MinLiquidity: Balance = MIN_LIQUIDITY;
-  pub const MinWeight: Balance = MIN_WEIGHT;
   pub const OracleBond: Balance = 50 * DOLLARS;
   pub const PmPalletId: PalletId = PalletId(*b"zge/pred");
   pub const ReportingPeriod: BlockNumber = DAYS;
   pub const SS58Prefix: u8 = 73;
-  pub const SwapsPalletId: PalletId = PalletId(*b"zge/swap");
   pub const TransactionByteFee: Balance = 1 * MILLICENTS;
   pub const ValidityBond: Balance = 50 * DOLLARS;
   pub const Version: RuntimeVersion = VERSION;
